requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
transformers
torch
faiss-cpu
sqlalchemy
requests
prometheus-client
redis
mkdocs-material
sentence-transformers